UserSettings ImageInput saved in database
after choosing an image when clicking on the image at the usersettings page you will be able to choose a picture from your file explorer and upload it or update it in your user after pressing the "save" button.
The updated image isn't loaded yet but only saved to the user.